Inline kvHandler to break import cycle
