+++ src/browser/toys.js
@@ -135,7 +135,71 @@ import { numberHandler } from '../inputHandlers/number.js';
 import { defaultHandler } from '../inputHandlers/default.js';
 import { dendriteStoryHandler } from '../inputHandlers/dendriteStory.js';
 import { dendritePageHandler } from '../inputHandlers/dendritePage.js';
-import { kvHandler } from '../inputHandlers/kv.js';
+import {
+  maybeRemoveNumber,
+  maybeRemoveDendrite,
+} from '../inputHandlers/removeElements.js';
+import { KV_CONTAINER_SELECTOR } from '../constants/selectors.js';
+import { hideAndDisable } from '../inputHandlers/inputState.js';
+
+export const ensureKeyValueInput = (container, textInput, dom) => {
+  let kvContainer = dom.querySelector(container, KV_CONTAINER_SELECTOR);
+  if (!kvContainer) {
+    kvContainer = dom.createElement('div');
+    dom.setClassName(kvContainer, KV_CONTAINER_SELECTOR.slice(1));
+    const nextSibling = dom.getNextSibling(textInput);
+    dom.insertBefore(container, kvContainer, nextSibling);
+  }
+
+  const rows = parseExistingRows(dom, textInput);
+  const disposers = [];
+
+  const render = createRenderer({
+    dom,
+    disposersArray: disposers,
+    container: kvContainer,
+    rows,
+    textInput,
+    syncHiddenField,
+  });
+
+  render();
+
+  const dispose = createDispose({
+    disposers,
+    dom,
+    container: kvContainer,
+    rows,
+  });
+  kvContainer._dispose = dispose;
+
+  return kvContainer;
+};
+
+/**
+ * Ensure a key/value input UI is displayed.
+ * @param {object} dom - DOM helper utilities.
+ * @param {HTMLElement} container - Container element for the inputs.
+ * @param {HTMLInputElement} textInput - The text input element.
+ * @returns {void}
+ */
+export function handleKVType(dom, container, textInput) {
+  maybeRemoveNumber(container, dom);
+  maybeRemoveDendrite(container, dom);
+  ensureKeyValueInput(container, textInput, dom);
+}
+
+/**
+ * Main handler for key/value input fields.
+ * @param {object} dom - DOM helper utilities.
+ * @param {HTMLElement} container - Container element for the inputs.
+ * @param {HTMLInputElement} textInput - The text input element.
+ * @returns {void}
+ */
+export function kvHandler(dom, container, textInput) {
+  hideAndDisable(textInput, dom);
+  handleKVType(dom, container, textInput);
+}
 
 const inputHandlersMap = {
   text: textHandler,
